main: Use RateLimitedClient to observe API rate limits
The Yammer REST API imposes rate limits on clients. If they are not
observed then all requests fail with HTTP response 429.

Swap out the http.Client with a RateLimitedClient, with parameters set
to allow 10 requests in 30 seconds.

Technically the Yammer rate limits are separated for each type of API
endpoint, but we only hit the /messages endpoint, which mandates 10
requests in 30 seconds.

See: https://developer.yammer.com/docs/rest-api-rate-limits